% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\name{check_font}
\alias{check_font}
\title{Check font is usable}
\usage{
check_font(fontname)
}
\arguments{
\item{fontname}{A possible}
}
\value{
The name of a font available to your R installation: Either the font given as fontname or the first available.
}
\description{
Check font is usable
}
\examples{
check_font("Cambria")

}
