% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_regular_table.R
\name{quick_regular_table}
\alias{quick_regular_table}
\title{Make a regular table from a dataframe with some easy formatting}
\usage{
quick_regular_table(x, ..., FONT = check_font("Cambria"),
  autofit = TRUE)
}
\arguments{
\item{x}{data frame}

\item{...}{additional arguments to flextable::regulartable}

\item{FONT}{A fontname, default Cambria or whatever is available}

\item{autofit}{Autofit column widths? (default TRUE)}
}
\value{
A regulartable object with selected font, bold header and (possibly) autofit widths
}
\description{
Make a regular table from a dataframe with some easy formatting
}
