% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1tidy.R
\name{t1tidy}
\alias{t1tidy}
\title{Tidy tableone output}
\usage{
t1tidy(x, newcol = "Variable", ...)
}
\arguments{
\item{x}{A tableone object}

\item{newcol}{name for the column of variables}

\item{...}{Other arguments passed to broom::fix_data_frame()}
}
\value{
A tibble
}
\description{
Tidy tableone output
}
